Create longestsubsequence.java
Given two strings text1 and text2, return the length of their longest common subsequence.

A subsequence is a sequence that appears in the same relative order but not necessarily contiguous. A common subsequence is a subsequence that appears in both strings.

Input Format

Two strings: text1 text2

Constraints

1 ≤ text1.length, text2.length ≤ 1000 text1 and text2 consist of only lowercase English letters

Output Format

An integer — the length of the longest common subsequence.

Sample Input 0

abcde
ace
Sample Output 0

3
Explanation 0

The longest common subsequence is "ace".